feat: add SNS verification
